+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- DAP: Format `sourceMap` correctly for both `codelldb` and `lldb`.
+  `codelldb` expects a map, while `lldb` expects a list of tuples.
+
 ## [3.6.5] - 2023-11-22
 
 ### Fixed